+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Vz Keymap extension will be documented in this file.
 ### [Unreleased]
 - Improved:
   - Behavior of Ctrl+L for Select words to find.
+  - Behavior of Escape key for Close find widget to do cancel selection as well.
   - Behavior of full size page scroll not to overshoot when the view is at the bottom of a document.
 
 ### [0.9.0] - 2020-02-25